+++ src/llm-providers/_common/register/getLlmToolsForCli.ts
@@ -3,7 +3,7 @@ import { EXECUTIONS_CACHE_DIRNAME } from '../../../config';
 import { EnvironmentMismatchError } from '../../../errors/EnvironmentMismatchError';
 import { FileCacheStorage } from '../../../storage/file-cache-storage/FileCacheStorage';
 import { $isRunningInNode } from '../../../utils/environment/$isRunningInNode';
-import { createLlmToolsFromEnv } from './createLlmToolsFromEnv';
+import { $provideLlmToolsFromEnv } from './$provideLlmToolsFromEnv';
 import { cacheLlmTools } from './utils/cache/cacheLlmTools';
 import { countTotalUsage } from './utils/count-total-usage/countTotalUsage';
 import type { LlmExecutionToolsWithTotalUsage } from './utils/count-total-usage/LlmExecutionToolsWithTotalUsage';
@@ -34,7 +34,7 @@ export function getLlmToolsForCli(options?: GetLlmToolsForCliOptions): LlmExecut
     return cacheLlmTools(
         countTotalUsage(
             //        <- Note: for example here we don`t want the [🌯]
-            createLlmToolsFromEnv(),
+            $provideLlmToolsFromEnv(),
         ),
         {
             storage: new FileCacheStorage(